Cleanup for logging and text handling
