+++ src/kenzy/extras.py
@@ -12,6 +12,20 @@ from . import __app_title__, __version__
 import json
 
 SSDP_DEVICE_TYPE = "urn:schemas-upnp-org:device:Kenzy-Core:1"
+ones = {
+    0: '', 1: 'one', 2: 'two', 3: 'three', 4: 'four', 5: 'five', 6: 'six',
+    7: 'seven', 8: 'eight', 9: 'nine', 10: 'ten', 11: 'eleven', 12: 'twelve',
+    13: 'thirteen', 14: 'fourteen', 15: 'fifteen', 16: 'sixteen',
+    17: 'seventeen', 18: 'eighteen', 19: 'nineteen'}
+
+tens = {
+    2: 'twenty', 3: 'thirty', 4: 'forty', 5: 'fifty', 6: 'sixty',
+    7: 'seventy', 8: 'eighty', 9: 'ninety'}
+
+illions = {
+    1: 'thousand', 2: 'million', 3: 'billion', 4: 'trillion', 5: 'quadrillion',
+    6: 'quintillion', 7: 'sextillion', 8: 'septillion', 9: 'octillion',
+    10: 'nonillion', 11: 'decillion'}
 
 
 class GenericCommand:
@@ -340,13 +354,13 @@ def parse_presentation_url(xml_data):
     return presentation_url
 
 
-def discover_ssdp_services(search_text=SSDP_DEVICE_TYPE):
+def discover_ssdp_services(search_text=SSDP_DEVICE_TYPE, timeout=45):
     logger = logging.getLogger("SSDP-CLT")
     ssdp_ip = "239.255.255.250"
     ssdp_port = 1900
 
     client_socket = socket.socket(socket.AF_INET, socket.SOCK_DGRAM)
-    client_socket.settimeout(5)
+    client_socket.settimeout(timeout)
     client_socket.setsockopt(socket.SOL_SOCKET, socket.SO_REUSEADDR, 1)
     client_socket.bind(("", ssdp_port))
 
@@ -386,3 +400,51 @@ def discover_ssdp_services(search_text=SSDP_DEVICE_TYPE):
         logger.critical("No UPNP Server identified via SSDP.")
 
     return url
+
+
+def strip_punctuation(text):
+    allowed = "abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789$.'"
+    text = "".join([str(x if x in allowed else " ") for x in [*text]]).replace("  ", " ")
+    return text
+
+
+def numbers_in_string(text):
+    return [i for i in text.replace(":", " ").split() if (i.strip("$?!.:;").replace(",", "").replace(".", "").isdigit())]
+
+
+def number_to_words(i):
+    """
+    Convert an integer in to it's word representation.
+
+    say_number(i: integer) -> string
+    """
+    if i < 0:
+        return _join('negative', _say_number_pos(-i))
+    if i == 0:
+        return 'zero'
+    return _say_number_pos(i)
+
+
+def _say_number_pos(i):
+    if i < 20:
+        return ones[i]
+    if i < 100:
+        return _join(tens[i // 10], ones[i % 10])
+    if i < 1000:
+        return _divide(i, 100, 'hundred')
+    for illions_number, illions_name in illions.items():
+        if i < 1000**(illions_number + 1):
+            break
+    return _divide(i, 1000**illions_number, illions_name)
+
+
+def _divide(dividend, divisor, magnitude):
+    return _join(
+        _say_number_pos(dividend // divisor),
+        magnitude,
+        _say_number_pos(dividend % divisor),
+    )
+
+
+def _join(*args):
+    return ' '.join(filter(bool, args))
